api/requirements.txt
flask==3.1.0
pyjwt==2.9.0
pika==1.3.2
psycopg2-binary==2.9.10
gunicorn==23.0.0
